requirements.txt
tensorflow-cpu
streamlit
opencv-python-headless
Pillow
numpy
gunicorn
requests